move profile to settings
