+++ app/src/features/settings/components/Profile/ManageTeams/TeamViewer/BillingDetails/InvoiceTable/InvoiceTable.css
@@ -15,10 +15,7 @@
   background-color: var(--component-background);
 }
 
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-thead
-  .ant-table-cell {
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-thead .ant-table-cell {
   padding: 0px 0px 8px 0px;
   font-size: 12px;
   line-height: 16px;
@@ -34,39 +31,20 @@
   visibility: hidden;
 }
 
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-thead
-  > tr
-  > th,
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-tbody
-  > tr
-  > td {
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-thead > tr > th,
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-tbody > tr > td {
   border-bottom: none;
 }
 
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-tbody
-  .ant-table-cell-row-hover {
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-tbody .ant-table-cell-row-hover {
   background-color: var(--component-background);
 }
 
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-tbody
-  .ant-table-row
-  td {
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-tbody .ant-table-row td {
   border-top: 1px solid #3e4044;
 }
 
-.invoice-details-table.ant-pro-table
-  .ant-table-wrapper
-  .ant-table-tbody
-  .ant-table-row
-  .ant-table-cell {
+.invoice-details-table.ant-pro-table .ant-table-wrapper .ant-table-tbody .ant-table-row .ant-table-cell {
   font-weight: 400;
   padding: 1.25rem 4px 1.25rem 2px;
 }
