+++ app/src/features/settings/components/Profile/ManageTeams/TeamViewer/MembersDetails/PublicInviteLink/index.css
@@ -1,11 +1,11 @@
 .invite-link-input-container {
-    margin-right: 8px;
+  margin-right: 8px;
 }
 
 .invite-link-input {
-    padding: 3px 8px !important;
+  padding: 3px 8px !important;
 }
 
 .invite-link-info-message {
-    margin-bottom: 16px;
+  margin-bottom: 16px;
 }
