+++ app/src/features/settings/components/SettingsPrimarySidebar/index.tsx
@@ -6,6 +6,7 @@ import { NavLink, useLocation, useNavigate } from "react-router-dom";
 import { MdOutlineDisplaySettings } from "@react-icons/all-files/md/MdOutlineDisplaySettings";
 import { RiBuildingLine } from "@react-icons/all-files/ri/RiBuildingLine";
 import { IoMdArrowBack } from "@react-icons/all-files/io/IoMdArrowBack";
+import { MdOutlineAccountBox } from "@react-icons/all-files/md/MdOutlineAccountBox";
 import { redirectToTraffic } from "utils/RedirectionUtils";
 import { isCompanyEmail } from "utils/FormattingHelper";
 import APP_CONSTANTS from "config/constants";
@@ -27,6 +28,18 @@ export const SettingsPrimarySidebar: React.FC = () => {
 
   const sidebarItems = useMemo(
     () => [
+      {
+        id: "my_accounts",
+        name: "Accounts",
+        icon: <MdOutlineAccountBox />,
+        children: [
+          {
+            id: "profile",
+            name: "Profile",
+            path: PATHS.SETTINGS.PROFILE.RELATIVE,
+          },
+        ],
+      },
       {
         id: "app_settings",
         name: "App",
